vosk-live-transcription/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-sock==0.7.0
Flask-SocketIO==5.5.1
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
propcache==0.3.1
pycparser==2.22
PyJWT==2.10.1
pyngrok==7.2.5
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
simple-websocket==1.1.0
srt==3.5.3
tqdm==4.67.1
twilio==9.5.2
urllib3==2.4.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
